pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stxtofr"
version = "0.1.0"
description = "Swedish transport facilities ingestion service"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.95,<0.111",
    "uvicorn>=0.21,<0.31",
    "requests>=2.28,<3",
    "SQLAlchemy>=1.4,<2.1",
    "alembic>=1.10,<1.14",
    "psycopg2-binary>=2.9,<3",
    "python-dotenv>=0.21,<2",
    "beautifulsoup4>=4.11,<5",
    "lxml>=4.9,<6",
    "geopy>=2.3,<3",
    "kafka-python>=2.0,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.2,<9",
    "httpx>=0.24,<0.29",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]

